Removes long select chains
