+++ src/components/schedule-page/table/schedule/schedule-parts/base-cell/base-cell-input.component.tsx
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import React from "react";
-import { CellManagementKeys } from "./base-cell.component";
+import { CellManagementKeys } from "./base-cell.models";
 
 export interface BaseCellInputOptions {
   className: string;
