+++ src/components/schedule-page/table/schedule/schedule-parts/shift-row.component.tsx
@@ -3,9 +3,10 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import React from "react";
 import { DataRow } from "../../../../../logic/schedule-logic/data-row";
-import { BaseRowComponent, BaseRowOptions } from "./base-row.component";
+import { BaseCellOptions } from "./base-cell/base-cell.models";
+import { BaseRowComponent } from "./base-row.component";
+import { BaseRowOptions } from "./base-row.models";
 import { ShiftCellComponent } from "./shift-cell/shift-cell.component";
-import { BaseCellOptions } from "./base-cell/base-cell.component";
 
 export interface ShiftRowOptions extends BaseRowOptions {
   dataRow: DataRow;
