+++ src/components/summarytable/summarytable-cell.component.tsx
@@ -2,14 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import React from "react";
+import { summaryCellDataCy, SummaryTableCellOptions } from "./summarytable-cell.models";
 
-export interface SummaryTableCellOptions {
-  value: number;
-}
-
-function SummaryTableCellF({ value }: SummaryTableCellOptions): JSX.Element {
+function SummaryTableCellF({ value, cellIndex }: SummaryTableCellOptions): JSX.Element {
   return (
-    <td className="summaryCell">
+    <td className="summaryCell" data-cy={summaryCellDataCy(cellIndex)}>
       <span>{value}</span>
     </td>
   );
