+++ src/components/summarytable/summarytable-row.component.tsx
@@ -2,19 +2,21 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import React from "react";
-import { SummaryTableCell } from "./summarytable-cell.component";
 import { ArrayHelper } from "../../helpers/array.helper";
+import { SummaryTableCell } from "./summarytable-cell.component";
+import { summaryRowDataCy, SummaryTableRowOptions } from "./summarytable-row.models";
 
-export interface SummaryTableRowOptions {
-  uuid: string;
-  data: number[];
-}
-
-export function SummaryTableRowF({ uuid, data }: SummaryTableRowOptions): JSX.Element {
+export function SummaryTableRowF({ uuid, data, rowIndex }: SummaryTableRowOptions): JSX.Element {
   return (
-    <tr className="row" id="summaryRow">
+    <tr className="row" id="summaryRow" data-cy={summaryRowDataCy(rowIndex)}>
       {data.map((cellData, cellIndex) => {
-        return <SummaryTableCell key={`${cellData}_${cellIndex}${uuid}}`} value={cellData} />;
+        return (
+          <SummaryTableCell
+            key={`${cellData}_${cellIndex}${uuid}}`}
+            value={cellData}
+            cellIndex={cellIndex}
+          />
+        );
       })}
     </tr>
   );
