+++ src/components/summarytable/summarytable-row.models.ts
@@ -1,0 +1,8 @@
+import { baseRowDataCy } from "../schedule-page/table/schedule/schedule-parts/base-row.models";
+
+export const summaryRowDataCy = (cellIndex: number): string => baseRowDataCy(cellIndex);
+export interface SummaryTableRowOptions {
+  uuid: string;
+  data: number[];
+  rowIndex: number;
+}
